Teuchos: More int/size_t warning cleanup

+++ packages/teuchos/core/src/Teuchos_RCPNode.cpp
@@ -329,7 +329,7 @@ int RCPNodeTracer::numActiveRCPNodes()
 {
   // This list always exists, no matter debug or not so just access it.
   TEUCHOS_TEST_FOR_EXCEPT(0==rcp_node_list());
-  return rcp_node_list()->size();
+  return static_cast<int>(rcp_node_list()->size());
 }
